+++ CHANGES.md
@@ -6,6 +6,8 @@
   - renamed `execute` schema component into `processRequest`
   - renamed `process` schema component into `processDescription`
   - renamed `results` schema component into `processResults`
+  - renamed `landingPage` schema component into `capabilities`
+  - renamed `confClasses` schema component into `conformanceDeclaration`
   - renamed `inline_response_200` schema component into `inlineValue`
   - renamed `inlineOrRefData` schema component into `inlineOrRefValue`
   - renamed `qualifiedInputValue` schema component into `qualifiedValue`